+++ packages/markdown/src/components/VoidElements/index.ts
@@ -1,3 +1,0 @@
-export * from './VoidElements';
-export { VoidElementsWarning } from './VoidElementsWarning';
-export { useVoidElements, VoidElementsContext } from './useVoidElements';